fix: Prevents mixed server and client code
