+++ packages/trpc/server/createContext.ts
@@ -3,7 +3,7 @@ import type { GetServerSidePropsContext, NextApiRequest, NextApiResponse } from 
 import type { Session } from "next-auth";
 import type { serverSideTranslations } from "next-i18next/serverSideTranslations";
 
-import { getLocaleFromRequest } from "@calcom/lib/i18n";
+import { getLocaleFromRequest } from "@calcom/lib/getLocaleFromRequest";
 import prisma from "@calcom/prisma";
 import type { SelectedCalendar, User as PrismaUser } from "@calcom/prisma/client";
